packages/adjutorix-agent/pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "adjutorix-agent"
version = "0.1.0"
description = "Adjutorix Local Agent Server (tools, governance, LLM routing)"
authors = [
  { name = "Adjutorix Team", email = "dev@adjutorix.local" }
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"

dependencies = [
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.27.0",
  "pydantic>=2.5.0",
  "jsonschema>=4.21.0",
  "requests>=2.31.0",
  "websockets>=12.0",
  "psutil>=5.9.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
  "gitpython>=3.1.40",
  "tree-sitter>=0.21.0",
  "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "black>=24.1.0",
  "ruff>=0.2.0",
  "mypy>=1.8.0",
  "types-PyYAML",
  "types-requests"
]

[project.scripts]
adjutorix-agent = "adjutorix_agent.server.app:main"

[tool.setuptools]
packages = ["adjutorix_agent"]

[tool.setuptools.package-data]
adjutorix_agent = [
  "llm/prompts/*.md",
  "llm/contracts/*.json"
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.ruff]
line-length = 88
select = ["E", "F", "W", "I", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.9"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q"
testpaths = ["tests"]
